Move `Token` and `Contract` to separate files
